backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy==2.0.36
pymysql==1.1.1
redis==5.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
